Fixed PlayerController stamina and dependence
Player Controller now does not drain stamina if the player is holding shift but is not moving. The script now gains stamina whenever the player is no moving, even if shift is held down.

Also, I made the playerController no longer throw errors if the gamemanager dependency doesn't exist, allowing it to work on its own.

Also ALSO, I made some edits to the test level and the temp enemy to fix the bug of bullets moving through walls (because, thats unfair)