Fixed bugs in previous commit
So, turns out the bugs I had were more a problem with the prefabs, so I went around and fixed those. The player can now switch guns and shoot according to its current gun's stats! The model also changes too!

However, whilst this solution works for now, I'm scared it could be heavy on the game because of it having to delete, then remake an item over and over. This'll be the fix for now until we can switch to meshes.